Some more cleaning up
