+++ src/features/index.ts
@@ -1,3 +1,4 @@
 export { InputMatcher } from "./InputMatcher";
 export { NoteLinksSync } from "./NoteLinksSync";
 export { NoteLoader } from "./NoteLoader";
+export { TimeTrackerTools } from "./TimeTrackingTools";
